requirements.txt
git+https://github.com/naya1503/pyrogram
tgcrypto
pytz
yt-dlp
youtube-search-python
motor
dnspython
GitPython
pykeyboard
bs4
Pillow
pymediainfo
requests
unidecode
attrify
ffmpeg-python
humanize
pyromod==1.1.0
html2text
removebg
aiofiles
aiohttp
gpytranslate
gtts
wget
telegraph
heroku3
openai
APScheduler
kynaylibs==1.9.2
psutil
httpx[http2]
uwuify
gunicorn
python-dotenv
uvloop
httpx